Mobile SDK: NFC scanning through SelfClient
